Split unit and integration into different files
